create lib folder and move utils inside
